requirements.txt
astronomer-cosmos>=1.4.3
dbt-core==1.8.1
dbt-snowflake==1.8.2